Refactor spaces app models to a package
