+++ spaces/models/__init__.py
@@ -1,0 +1,16 @@
+from .building import Building
+from .location import Location
+from .real_estate import RealEstate
+from .service_sector import ServiceSector
+from .space import Space
+from .unit import Unit, UnitGroup
+
+__all__ = [
+    "Building",
+    "Location",
+    "RealEstate",
+    "ServiceSector",
+    "Space",
+    "Unit",
+    "UnitGroup",
+]
